requirements.txt
Django==1.10.5
django-compressor==2.11
djangorestframework==3.5.3
